refactor: remove AmplitudeGenerator.write_to_file
The idea is to have all writing functionality slowly migrated to the
expertsystem.io model and have the rest of the framework produce python
classes that can then be serialized through the io module.
