+++ tests/unit/amplitude/test_yaml_canonical.py
@@ -6,16 +6,16 @@ import pytest
 
 import yaml
 
+from expertsystem.amplitude import write_model
+
 
 SCRIPT_PATH = dirname(realpath(__file__))
 
 
 @pytest.fixture(scope="module")
-def imported_dict(jpsi_to_gamma_pi_pi_canonical_amplitude_generator):
+def imported_dict(jpsi_to_gamma_pi_pi_canonical_amplitude_model):
     output_filename = "JPsiToGammaPi0Pi0_cano_recipe.yml"
-    jpsi_to_gamma_pi_pi_canonical_amplitude_generator.write_to_file(
-        output_filename
-    )
+    write_model(jpsi_to_gamma_pi_pi_canonical_amplitude_model, output_filename)
     with open(output_filename, "rb") as input_file:
         loaded_dict = yaml.load(input_file, Loader=yaml.FullLoader)
     return loaded_dict
@@ -27,11 +27,12 @@ def equalize_dict(input_dict):
 
 
 def test_not_implemented_writer(
-    jpsi_to_gamma_pi_pi_canonical_amplitude_generator,
+    jpsi_to_gamma_pi_pi_canonical_amplitude_model,
 ):
     with pytest.raises(NotImplementedError):
-        jpsi_to_gamma_pi_pi_canonical_amplitude_generator.write_to_file(
-            "JPsiToGammaPi0Pi0.csv"
+        write_model(
+            jpsi_to_gamma_pi_pi_canonical_amplitude_model,
+            "JPsiToGammaPi0Pi0.csv",
         )
